requirements.txt
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
joblib==1.5.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2